+++ fs/text/sgr/proof.f.ts
@@ -1,6 +1,6 @@
 import { fgRed, reset, createConsoleText, backspace, csiWrite } from './module.f.ts'
-import { virtual, emptyState } from '../../effects/node/virtual/module.f.ts'
-import { defaultNodeProgramOptions, type NodeProgramOptions } from '../../effects/node/module.f.ts'
+import { virtual, emptyState, defaultNodeProgramOptions } from '../../effects/node/virtual/module.f.ts'
+import type { NodeProgramOptions } from '../../effects/node/module.f.ts'
 
 const makeOptions = (isTTY: boolean): NodeProgramOptions =>
     ({ ...defaultNodeProgramOptions, std: { stdout: { isTTY }, stderr: { isTTY } } })
